+++ metrics/src/metrics-logger.js
@@ -1,0 +1,13 @@
+export class MetricsLogger {
+  constructor() {
+    this.entries = [];
+  }
+
+  log(entry) {
+    this.entries.push(entry);
+  }
+
+  getEntries() {
+    return [...this.entries];
+  }
+}